+++ src/utils/validation.ts
@@ -1,5 +1,6 @@
 import * as vscode from 'vscode';
 import { Structure } from '../types';
+import { promptShowInvalidReason } from './promptUtils';
 
 export function isValidName(name: string): boolean {
     const forbidden = /[\\\/:\*\?"<>\|]/;
@@ -54,40 +55,59 @@ export async function isValidStructure(structure: Structure): Promise<boolean> {
     const structureOptionals = structure.optionals || [];
     const structureStructure = structure.structure;
 
+    let errorMessage = '';
+
     // Exit if structure doesn't have any files
     if (!structureStructure) {
-        vscode.window.showErrorMessage('This structure doesn\'t have any files. Please update your structure', { modal: true });
-        return false;
+        errorMessage += '\n    - The structure doesn\'t contain any files';
     }
 
     // Exit if duplicate variable
     const uniqueVariables = new Set<string>();
+    const duplicateVariables = new Set<string>();
     for (const variable of structureVariables) {
-        if (uniqueVariables.has(variable.varName)) {
-            vscode.window.showErrorMessage(`Duplicate variable '${variable.varName}'. Please update your structure`, { modal: true });
-            return false;
+        const name = variable.varName;
+        if (uniqueVariables.has(name) && !duplicateVariables.has(name)) {
+            errorMessage += `\n    - Duplicate variable '${name}'`;
+            duplicateVariables.add(name);
         }
-        uniqueVariables.add(variable.varName);
+        uniqueVariables.add(name);
     }
 
     // Exit if duplicate optional
     const uniqueOptionals = new Set<string>();
+    const duplicateOptionals = new Set<string>();
     for (const optional of structureOptionals) {
-        if (uniqueOptionals.has(optional.optName)) {
-            vscode.window.showErrorMessage(`Duplicate optional '${optional.optName}'. Please update your structure`, { modal: true });
-            return false;
+        const name = optional.optName;
+        if (uniqueOptionals.has(name) && !duplicateOptionals.has(name)) {
+            errorMessage += `\n    - Duplicate optional '${name}'`;
+            duplicateOptionals.add(name);
         }
-        uniqueOptionals.add(optional.optName);
+        uniqueOptionals.add(name);
+    }
+
+    // Exit if in both variables and optionals
+    for (const name of uniqueOptionals) {
+        if (uniqueVariables.has(name)) {
+            errorMessage += `\n    - Name '${name}' is in both variables and optionals`;
+        }
     }
 
     // Exit if duplicate file
     const uniqueFiles = new Set<string>();
+    const duplicateFiles = new Set<string>();
     for (const file of structureStructure) {
-        if (uniqueFiles.has(file.fileName)) {
-            vscode.window.showErrorMessage(`Duplicate file '${file.fileName}'. Please update your structure`, { modal: true });
-            return false;
+        const name = file.fileName;
+        if (uniqueFiles.has(name) && !duplicateFiles.has(name)) {
+            errorMessage += `\n    - Duplicate file '${name}'`;
+            duplicateFiles.add(name);
         }
-        uniqueFiles.add(file.fileName);
+        uniqueFiles.add(name);
+    }
+
+    if (errorMessage) {
+        promptShowInvalidReason(errorMessage);
+        return false;
     }
 
     return true;